+++ src/suews/src/suews_type_stebbs.f95
@@ -70,10 +70,10 @@ module module_type_stebbs
       REAL(KIND(1D0)) :: MaxHeatingPower = 0.0D0 ! Maximum power demand of heating system [W]
       REAL(KIND(1D0)) :: WaterTankWaterVolume = 0.0D0 ! Volume of water in hot water tank [m3]
       REAL(KIND(1D0)) :: MaximumHotWaterHeatingPower = 0.0D0 ! Maximum power demand of water heating system [W]
-      REAL(KIND(1D0)), DIMENSION(0:143, 2) :: HeatingSetpointTemperature = 0.0D0 ! Heating setpoint temperature [degC]
-      REAL(KIND(1D0)), DIMENSION(0:143, 2) :: CoolingSetpointTemperature = 0.0D0 ! Cooling setpoint temperature [degC]
-      REAL(KIND(1D0)), DIMENSION(0:143, 2) :: OccupantsProfile = 0.0D0 ! diurnal profiles of occupants presence
-      REAL(KIND(1D0)), DIMENSION(0:143, 2) :: ApplianceProfile = 0.0D0 ! diurnal profiles of appliance energy power
+      REAL(KIND(1D0)) :: HeatingSetpointTemperature = 0.0D0 ! Heating setpoint temperature [degC]
+      REAL(KIND(1D0)) :: CoolingSetpointTemperature = 0.0D0 ! Cooling setpoint temperature [degC]
+      REAL(KIND(1D0)), DIMENSION(0:143, 2) :: MetabolismProfile = 0.0D0 ! diurnal profiles of occupant metabolic rate [W]
+      REAL(KIND(1D0)), DIMENSION(0:143, 2) :: ApplianceProfile = 0.0D0 ! diurnal profiles of appliance energy power [W]
       ! flag for iteration safety - YES - as we this should be updated every iteration
       LOGICAL :: iter_safe = .TRUE.
    END TYPE BUILDING_ARCHETYPE_PRM
@@ -92,9 +92,8 @@ module module_type_stebbs
       REAL(KIND(1D0)) :: ExternalGroundConductivity = 0.0D0
       REAL(KIND(1D0)) :: IndoorAirDensity = 0.0D0 ! Density of indoor air [kg m-3]
       REAL(KIND(1D0)) :: IndoorAirCp = 0.0D0 ! Specific heat capacity of indoor air [J kg-1 K-1]
-      REAL(KIND(1D0)) :: MetabolicRate = 0.0D0 ! Metabolic rate of building occupants [W]
+      REAL(KIND(1D0)) :: MetabolismThreshold = 0.0D0 ! Threshold of Metabolic rate of each occupancy for active or inactive [W]
       REAL(KIND(1D0)) :: LatentSensibleRatio = 0.0D0 ! Latent-to-sensible ratio of metabolic energy release of occupants [-]
-      REAL(KIND(1D0)) :: ApplianceRating = 0.0D0 ! Power demand of single appliance [W]
       REAL(KIND(1D0)) :: HeatingSystemEfficiency = 0.0D0 ! Efficiency of space heating system [-]
       REAL(KIND(1D0)) :: MaxCoolingPower = 0.0D0 ! Maximum power demand of cooling system [W]
       REAL(KIND(1D0)) :: CoolingSystemCOP = 0.0D0 ! Coefficient of performance of cooling system [-]
@@ -207,11 +206,10 @@ module module_type_stebbs
       REAL(KIND(1D0)) :: roofAbsorbtivity = 0.0D0
       REAL(KIND(1D0)) :: roofReflectivity = 0.0D0
       REAL(KIND(1D0)) :: occupants = 0.0D0
-      REAL(KIND(1D0)) :: frac_occupants = 0.0D0
       REAL(KIND(1D0)) :: metabolic_rate = 0.0D0
+      REAL(KIND(1D0)) :: metabolism_threshold = 0.0D0
       REAL(KIND(1D0)) :: ratio_metabolic_latent_sensible = 0.0D0
       REAL(KIND(1D0)) :: appliance_power_rating = 0.0D0
-      REAL(KIND(1D0)) :: frac_appliance = 0.0D0
       REAL(KIND(1D0)) :: frac_hotwater = 0.0D0
       REAL(KIND(1D0)) :: maxheatingpower_air = 0.0D0
       REAL(KIND(1D0)) :: heating_efficiency_air = 0.0D0
@@ -308,8 +306,6 @@ module module_type_stebbs
       REAL(KIND(1D0)), DIMENSION(6) :: Cp = 0.0D0
       REAL(KIND(1D0)), DIMENSION(7) :: emis = 0.0D0
       REAL(KIND(1D0)), DIMENSION(3) :: wiTAR, waTAR, roofTAR = 0.0D0
-      !REAL(KIND(1D0)), DIMENSION(6) :: viewFactors = 0.0D0
-      REAL(KIND(1D0)), DIMENSION(3) :: occupantData = 0.0D0
       REAL(KIND(1D0)), DIMENSION(3) :: HTsAverage, HWTsAverage = 0.0D0
       REAL(KIND(1D0)), DIMENSION(3) :: HWPowerAverage = 0.0D0
       REAL(KIND(1D0)), DIMENSION(44) :: EnergyExchanges = 0.0D0